+++ UnitTests~/ReactiveComponent/WD/ReactiveComponentWriteDefaultsTest.cs
@@ -27,7 +27,7 @@ public class ReactiveComponentWriteDefaultsTest : TestBase
 
         AvatarProcessor.ProcessAvatar(root);
 
-        var fx = findFxLayer(root, "RC MA Responsive: GameObject");
+        var fx = findFxLayer(root, "MA Responsive: GameObject");
         Assert.AreEqual(
             controllerPath.StartsWith("rc_wd_on"),
             fx.stateMachine.defaultState.writeDefaultValues